feed/requirements.txt
python-dotenv
pyfiglet
numpy
requests
flask
flask-classy
bson
selenium
pymongo
slimit
beautifulsoup4
elasticsearch
pyyaml
docker
influxdb
hazelcast-python-client
kafka-python
